Major improvements made to search engine mechanics

+++ materialsearchbarlibary/src/main/java/com/eudycontreras/materialsearchbarlibary/fragments/SearchResultFragment.kt
@@ -110,7 +110,7 @@ class SearchResultFragment : Fragment() {
         createSpringAnimation(0f, 1f, duration, distance, friction, tension,
                 { value ->
                     searchResultContainerBackground.animate()
-                            .scaleY(value.toFloat())
+                            .scaleY(value)
                             .alpha(1f)
                             .setDuration(0)
                             .start()
